AshkanRafie/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-js-asset==3.0.1
jalali_core==1.0.0
jdatetime==5.2.0
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
Unidecode==1.3.8
